+++ README.md
@@ -3,16 +3,16 @@
 ## 기능 구현 목록 
 
 ### 카드 추가 뷰 구현
-- [ ] 카드 번호 입력 필드 구현 
+- [x] 카드 번호 입력 필드 구현 
   - [x] 초기 상태 : 힌트로 "0000 - 0000 - 0000 - 0000" 표시
   - [x] 숫자만 입력 가능하도록 키보드 타입 설정
-  - 입력 시에, 4자리를 기준으로 하이픈(-) 자동 삽입
+  - [x] 입력 시에, 4자리를 기준으로 하이픈(-) 자동 삽입
   - [x] 16자를 초과해 입력할 수 없음
 
-- [ ] 만료일 입력 필드 구현 
+- [x] 만료일 입력 필드 구현 
   - [x] 초기 상태 : 힌트로 "MM/YY" 표시
   - [x] 숫자만 입력 가능하도록 키보드 타입 설정
-  - MM/YY 형식으로 자동 슬래시(/) 추가
+  - [x] MM/YY 형식으로 자동 슬래시(/) 추가
   - [x] 4자를 초과해 입력할 수 없음
 
 - [x] 카드 소유자 이름 입력 필드 구현
@@ -21,10 +21,10 @@
   - [x] 30자를 초과해 입력할 수 없다. 
   - [x] 하단에 현재 입력한 글자수 출력
 
-- [ ] 비밀번호 입력 필드 구현
-  - 초기 상태 : 힌트로 "0000" 표시 
-  - 입력 내용이 보이지 않도록 마스킹 처리
-  - 4자를 초과해 입력할 수 없다
+- [x] 비밀번호 입력 필드 구현
+  - [x] 초기 상태 : 힌트로 "0000" 표시 
+  - [x] 입력 내용이 보이지 않도록 마스킹 처리
+  - [x] 4자를 초과해 입력할 수 없다
 
 - [ ] 유효성 검사 및 UI 상호작용
   - 필수 입력 필드(카드 번호, 만료일, 비밀번호)가 유효성 검사를 통과했을 때만 상단 확인 버튼 활성화
