+++ app/src/main/java/woowacourse/payments/card/register/component/CardPasswordTextField.kt
@@ -1,0 +1,40 @@
+package woowacourse.payments.card.register.component
+
+import androidx.compose.foundation.text.KeyboardOptions
+import androidx.compose.material3.OutlinedTextField
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.input.KeyboardType
+import androidx.compose.ui.text.input.PasswordVisualTransformation
+import androidx.compose.ui.tooling.preview.Preview
+
+@Preview
+@Composable
+fun CardPasswordTextField(modifier: Modifier = Modifier) {
+    var password by remember { mutableStateOf("") }
+    val numericRegex = Regex("[^0-9]")
+
+    OutlinedTextField(
+        value = password,
+        onValueChange = {
+            val stripped = numericRegex.replace(it, "")
+            password = if (stripped.length <= 4) {
+                stripped
+            } else {
+                stripped.substring(0, 4)
+            }
+        },
+        label = { Text("비밀번호") },
+        placeholder = { Text("0000") },
+        keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Number),
+        visualTransformation = PasswordVisualTransformation(),
+        singleLine = true,
+        modifier = modifier,
+    )
+}
+
