+++ app/src/main/java/woowacourse/payments/card/register/component/NewCardTopBar.kt
@@ -1,0 +1,41 @@
+package woowacourse.payments.card.register.component
+
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.automirrored.filled.ArrowBack
+import androidx.compose.material.icons.filled.Check
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.Text
+import androidx.compose.material3.TopAppBar
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun NewCardTopBar(
+    onBackClick: () -> Unit,
+    onSaveClick: () -> Unit,
+    modifier: Modifier = Modifier,
+) {
+    TopAppBar(
+        title = { Text("카드 추가") },
+        navigationIcon = {
+            IconButton(onClick = { onBackClick() }) {
+                Icon(
+                    imageVector = Icons.AutoMirrored.Filled.ArrowBack,
+                    contentDescription = "뒤로 가기",
+                )
+            }
+        },
+        actions = {
+            IconButton(onClick = { onSaveClick() }) {
+                Icon(
+                    imageVector = Icons.Filled.Check,
+                    contentDescription = "완료",
+                )
+            }
+        },
+        modifier = modifier,
+    )
+}
